requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.7
    # via django
atomicwrites==1.4.0
    # via pytest
attrs==21.2.0
    # via pytest
certifi==2021.5.30
    # via requests
chardet==4.0.0
    # via requests
colorama==0.4.4
    # via pytest
django-filter==2.4.0
    # via -r requirements.in
django==3.0.5
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==4.7.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
idna==2.10
    # via requests
importlib-metadata==4.5.0
    # via
    #   pluggy
    #   pytest
iniconfig==1.1.1
    # via pytest
packaging==20.9
    # via pytest
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pyjwt==2.1.0
    # via djangorestframework-simplejwt
pyparsing==2.4.7
    # via packaging
pytest-django==4.4.0
    # via -r requirements.in
pytest==6.2.4
    # via pytest-django
python-dotenv==0.17.1
    # via -r requirements.in
pytz==2019.3
    # via django
requests==2.25.1
    # via -r requirements.in
sqlparse==0.3.1
    # via django
toml==0.10.2
    # via pytest
typing-extensions==3.10.0.0
    # via importlib-metadata
urllib3==1.26.5
    # via requests
zipp==3.4.1
    # via importlib-metadata
